+++ equinity-historical-data/src/main/java/uk/gov/justice/laa/crime/equinity/historicaldata/repository/criteria/CrmFormDetailsCriteria.java
@@ -4,11 +4,8 @@ import jakarta.annotation.Nullable;
 import lombok.NoArgsConstructor;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.data.jpa.domain.Specification;
-import uk.gov.justice.laa.crime.equinity.historicaldata.repository.criteria.input.CrmFormDetailsCriteriaDTO;
 import uk.gov.justice.laa.crime.equinity.historicaldata.model.CrmFormModelInterface;
-
-import java.util.Arrays;
-import java.util.List;
+import uk.gov.justice.laa.crime.equinity.historicaldata.repository.criteria.input.CrmFormDetailsCriteriaDTO;
 
 @Configuration
 @NoArgsConstructor
@@ -17,7 +14,6 @@ public class CrmFormDetailsCriteria {
         return Specification
             .where(byUsn(crmFormDetailsCriteriaDTO.usn())
                 .and(byType(crmFormDetailsCriteriaDTO.type()))
-                .and(byProfileAcceptedTypes(crmFormDetailsCriteriaDTO.profileAcceptedTypes()))
             );
     }
 
@@ -29,13 +25,5 @@ public class CrmFormDetailsCriteria {
     private Specification<CrmFormModelInterface> byType(@Nullable Integer type) {
         return (root, query, criteriaBuilder)
                 -> type == null ? null : criteriaBuilder.equal(root.get("typeId"), type);
-    }
-
-    private Specification<CrmFormModelInterface> byProfileAcceptedTypes(@Nullable String types) {
-        if (types == null) return null;
-
-        List<String> convertedTypes = Arrays.asList(types.replace(" ", "").split(",", -1));
-        return (root, query, criteriaBuilder)
-                -> root.get("typeId").in(convertedTypes);
     }
 }
