+++ equinity-historical-data/src/main/java/uk/gov/justice/laa/crime/equinity/historicaldata/repository/criteria/input/CrmFormDetailsCriteriaDTO.java
@@ -2,6 +2,7 @@ package uk.gov.justice.laa.crime.equinity.historicaldata.repository.criteria.inp
 
 import jakarta.annotation.Nullable;
 import org.springframework.dao.InvalidDataAccessApiUsageException;
+import uk.gov.justice.laa.crime.equinity.historicaldata.util.ProfileAcceptedTypesUtil;
 
 public record CrmFormDetailsCriteriaDTO(
         Long usn,
@@ -12,6 +13,8 @@ public record CrmFormDetailsCriteriaDTO(
         if (usn == null) {
             throw new InvalidDataAccessApiUsageException("Expected USN not be null");
         }
+
+        ProfileAcceptedTypesUtil.checkTypeIsAcceptedByProfile(type, profileAcceptedTypes);
     }
 
     @Override
