+++ equinity-historical-data/src/main/java/uk/gov/justice/laa/crime/equinity/historicaldata/util/ProfileAcceptedTypesUtil.java
@@ -1,0 +1,18 @@
+package uk.gov.justice.laa.crime.equinity.historicaldata.util;
+
+import jakarta.annotation.Nullable;
+import lombok.experimental.UtilityClass;
+import uk.gov.justice.laa.crime.equinity.historicaldata.exception.UnauthorizedUserProfileException;
+
+@UtilityClass
+public class ProfileAcceptedTypesUtil {
+    public static void checkTypeIsAcceptedByProfile(Integer requiredType, @Nullable String types) throws UnauthorizedUserProfileException {
+        if (types == null) return;
+
+        if (!types.contains(String.valueOf(requiredType))) {
+            throw new UnauthorizedUserProfileException(
+                    String.format("Unauthorized. User profile does not have privileges to access requested report type [%d] ", requiredType)
+            );
+        }
+    }
+}
